Added internal cache for definitions
During the save process, getFieldDefinitionFromJson is called multiple times. In vanilla Pimcore at least twice for each definition - during save and when pushing data to backend search. When AOS plugin is installed - that's another one.

This change caches the result of getFieldDefinitionFromJson in static array. As the gain during regular save is minimal, the real benefit can be seen when saving a lot of objects in one go.

Here's the comparision for importing data from file (~100 objects with classification store). Before:
https://blackfire.io/profiles/43a7d2bf-0adc-45e9-9bf0-530a8c3e564a/graph
After:
https://blackfire.io/profiles/0be9a7e9-162a-40fa-90f7-247605a5895f/graph